Big commit. Added data.json and almost completed learn and train functionality for both black and white. Also updated import fen function with more efficient version.
This commit includes a user colour and colour to move variables, which change with the flip function and navigations
In the UI, added learn and train buttons and added safeguards for learn, so that there is no wrong data saved in the AI.
Also added file handling for data.json.
Finally in AI class, used the moves list to train the AI rather than recording the moves in realtime.